build ZXBox from Ubuntu 20.04 desktop

+++ install_zxbox_ubuntu2004.sh
@@ -2,7 +2,6 @@
 SCRIPTDIR=$(cd `dirname $0` && pwd)
 
 apt-get update
-apt-get install -y xinit
 
 # enable autologin
 mv /etc/gdm3/custom.conf /etc/gdm3/custom.conf.bak
@@ -12,6 +11,7 @@ echo "AutomaticLogin = vagrant" >> /etc/gdm3/custom.conf
 
 
 apt-get install -y dosbox wine libgles2-mesa
+dpkg --add-architecture i386 && sudo apt-get update && sudo apt-get install wine32
 
 apt-get install -y fuse-emulator-sdl
 
@@ -20,6 +20,20 @@ apt-get install -y fuse-emulator-sdl
 mv $SCRIPTDIR/emul/usp $SCRIPTDIR/emul/usp.old
 bash $SCRIPTDIR/emul/build_usp_debian-ubuntu.sh
 
+# create shortcuts to lunch emulators
+ln -s $SCRIPTDIR /home/vagrant/Desktop
+cp $SCRIPTDIR/usp.desktop /home/vagrant/Desktop/
 
-# after install xinit boot into gui logon
-reboot
+
+if [ -e $FROM_UBUNTUS_ERVER ]; then
+  # make ubuntu desktop from ubuntu server
+  apt-get install -y xinit
+  apt-get install -y gnome-panel gnome-tweaks
+  # add Nemo default file manager in Ubuntu
+  apt-get install -y nemo
+  xdg-mime default nemo.desktop inode/directory application/x-gnome-saved-search
+  gsettings set org.gnome.desktop.background show-desktop-icons false
+  gsettings set org.nemo.desktop show-desktop-icons true
+  # after install xinit boot into gui logon
+  reboot
+fi
